Move existing NER source to bilstm_crf dir
